fixed server startup, session handling, and web API request parameter parsing

+++ Startup.cs
@@ -24,6 +24,9 @@ namespace Websilk
             //load application-wide memory store
             Server server = new Server();
 
+            //use session
+            app.UseSession();
+
             //handle static files
             var options = new StaticFileOptions {ContentTypeProvider = new FileExtensionContentTypeProvider()};
             app.UseStaticFiles(options);
@@ -33,15 +36,8 @@ namespace Websilk
             errOptions.SourceCodeLineCount = 10;
             app.UseDeveloperExceptionPage();
 
-            //use session
-            app.UseSession();
-
             //get server info from config.json
-            if (!File.Exists(Path.GetFullPath("config.json")))
-            {
-                //generate config.json
-                createConfig();
-            }
+            checkForConfig();
 
             var config = new ConfigurationBuilder()
                 .AddJsonFile(server.MapPath("config.json"))
@@ -50,8 +46,6 @@ namespace Websilk
             server.sqlActive = config.GetSection("Data:Active").Value;
             server.sqlConnection = config.GetSection("Data:" + server.sqlActive).Value;
             server.https = config.GetSection("https").Value.ToLower() == "true" ? true : false;
-            server.encryption.salt = config.GetSection("Encryption:Salt").Value;
-            server.encryption.spliceIndex = int.Parse(config.GetSection("Encryption:Splice").Value);
 
             var isdev = false;
             switch (config.GetSection("Environment").Value)
@@ -69,25 +63,8 @@ namespace Websilk
             }
 
             //configure server security
-            if(server.encryption.salt == "")
-            {
-                //if no salt exists in config.json, create a new one
-                var Util = new Utility.Util();
-                server.encryption.salt = createSalt(Util);
-
-                //reset all passwords in the database
-                var Sql = new Sql(server, Util);
-                Sql.ExecuteNonQuery("EXEC ResetAllPasswords");
-                server.encryption.reset = true;
-            }else
-            {
-                //check to see if admin account has a password
-                var Util = new Utility.Util();
-                var Sql = new Sql(server, Util);
-                if((int)Sql.ExecuteScalar("EXEC HasPassword @userId=1") == 0){
-                    server.encryption.reset = true;
-                }
-            }
+            server.bcrypt_workfactor = int.Parse(config.GetSection("Encryption:bcrypt_work_factor").Value);
+            server.CheckAdminPassword();
 
             //run Websilk application
             app.Run(async (context) =>
@@ -135,7 +112,6 @@ namespace Websilk
                                 form = await context.Request.ReadFormAsync();
                             }
                         }
-
                         
                         if (cleanNamespace(paths))
                         {
@@ -197,38 +173,13 @@ namespace Websilk
             return true;
         }
 
-        private string createSalt(Utility.Util Util)
-        {
-            //no salt has been generated yet. generate salt & save to config.json
-            var encryption = new Utility.Encryption(Util);
-            var configPath = Path.GetFullPath("config.json");
-            var salt = encryption.getNewCSPRNG();
-            //open config.json as raw file
-            var f = File.ReadAllText(configPath);
-            if (f != "")
-            {
-                //parse config.json & insert salt manually
-                var i = new int[3];
-                i[0] = f.IndexOf("\"salt\"");
-                if (i[0] > 0)
-                {
-                    i[1] = f.IndexOf("\"\"", i[0] + 6);
-                    if (i[1] > i[0])
-                    {
-                        //inject salt into config.json
-                        f = f.Substring(0, i[1] + 1) + salt + f.Substring(i[1] + 1);
-                        File.WriteAllText(configPath, f);
-                    }
-                }
-            }
-            return salt;
-        }
-
-        private void createConfig()
+        private void checkForConfig()
         {
             //generate a new config.json file
             var configPath = Path.GetFullPath("project.json").Replace("project.json","config.json");
-            File.WriteAllText(configPath,
+            if (!File.Exists(configPath))
+            {
+                File.WriteAllText(configPath,
                 "{\n" +
                 "    \"environment\": \"development\",\n" +
                 "    \"https\": false,\n" +
@@ -237,12 +188,12 @@ namespace Websilk
                 "        \"SqlServerDev\": \"server=.\\\\SQL2014; database=WebsilkDev; user=WebsilkDev; password=development\",\n" +
                 "        \"SqlServerTrusted\": \"server=.\\\\SQL2014; database=WebsilkDev; Trusted_Connection=true\"\n" +
                 "    },\n" +
-                "    \"encryption\": {\n" +
-                "        \"salt\": \"\",\n" +
-                "        \"splice\": 16\n" +
+                "    \"encryption\":{\n" +
+                "        \"bcrypt_work_factor\":\"10\"\n" +
                 "    }\n" +
                 "}\n"
                 );
+            }
         }
     }
 }
